Optimized fract-ol code and fixed phoenix fractal value initialization

+++ fract-ol/srcs/mouse_events.c
@@ -6,7 +6,7 @@
 /*   By: maaugust <maaugust@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/30 15:45:23 by maaugust          #+#    #+#             */
-/*   Updated: 2025/07/01 03:33:15 by maaugust         ###   ########.fr       */
+/*   Updated: 2025/07/01 13:12:18 by maaugust         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ static void	screen_zoom(int button, int x, int y, t_fractal *frac)
 	double	prev_off_re;
 	double	prev_off_im;
 
-	prev_off_re = ((XMAX - XMIN) * x / frac->width + XMIN) * frac->zoom
+	prev_off_re = ((XMAX - XMIN) * (double)x / frac->width + XMIN) * frac->zoom
 		+ frac->off_x;
-	prev_off_im = ((YMAX - YMIN) * y / frac->height + YMIN) * frac->zoom
+	prev_off_im = ((YMAX - YMIN) * (double)y / frac->height + YMIN) * frac->zoom
 		+ frac->off_y;
 	if (button == MOUSE_SCROLL_UP)
 		frac->zoom *= ZOOM_SCALE;
